refactor: modularize files for easy of use
