+++ atoma-p2p/Cargo.toml
@@ -22,6 +22,7 @@ libp2p = { workspace = true, features = [
 	"tcp",
 	"yamux",
 	"noise",
+	"metrics",
 ] }
 fastcrypto = { workspace = true }
 flume = { workspace = true }
