+++ atoma-p2p/src/lib.rs
@@ -1,8 +1,13 @@
 pub mod config;
+pub mod errors;
 pub mod metrics;
 pub mod service;
 pub mod timer;
 pub mod types;
+pub mod utils;
+
+#[cfg(test)]
+mod tests;
 
 pub use config::AtomaP2pNodeConfig;
 pub use service::AtomaP2pNode;
