+++ deny.toml
@@ -24,6 +24,8 @@ ignore = [
 	"RUSTSEC-2024-0320",
 	# allow unmaintained proc-macro-error used in transitive dependencies
 	"RUSTSEC-2024-0370",
+	# ring is unmaintained but is a critical dependency through sui-sdk and libp2p
+	"RUSTSEC-2025-0007",
 ]
 # Threshold for security vulnerabilities, any vulnerability with a CVSS score
 # lower than the range specified will be ignored. Note that ignored advisories
